% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2.R
\name{align_sequences}
\alias{align_sequences}
\title{Align Sequences Using Minimap2}
\usage{
align_sequences(aligner, query_seq, query_name = "", cs = FALSE, MD = FALSE)
}
\arguments{
\item{aligner}{A Minimap2Aligner object}

\item{query_seq}{Query sequence to align}

\item{query_name}{Name of the query sequence}

\item{cs}{Whether to output the cs tag}

\item{MD}{Whether to output the MD tag}
}
\value{
A list of alignments
}
\description{
Align Sequences Using Minimap2
}
