% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\docType{class}
\name{AlignmentRecords-class}
\alias{AlignmentRecords-class}
\title{AlignmentRecords Class}
\description{
AlignmentRecords Class
}
\section{Slots}{

\describe{
\item{\code{header}}{Character vector storing SAM header lines}

\item{\code{records}}{tibble containing alignment records}
}}

