% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{AlignmentRecords}
\alias{AlignmentRecords}
\title{Create a new AlignmentRecords object}
\usage{
AlignmentRecords(
  header = character(),
  records = tibble::tibble(qname = character(), flag = integer(), rname = character(),
    pos = integer(), mapq = integer(), cigar = character(), rnext = character(), pnext =
    integer(), tlen = integer(), seq = character(), qual = character(), tags =
    character())
)
}
\arguments{
\item{header}{Character vector of SAM header lines}

\item{records}{tibble containing alignment records}
}
\value{
An AlignmentRecords object
}
\description{
Create a new AlignmentRecords object
}
