% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build a minimap2 index from a reference sequence}
\usage{
build_index(reference_file, preset = "map-ont", n_threads = 3)
}
\arguments{
\item{reference_file}{Path to the reference sequence file (FASTA format)}

\item{preset}{Character string specifying the preset to use. Default is "map-ont".
Available presets include:
\itemize{
\item "map-ont" - Oxford Nanopore read mapping
\item "map-pb" - PacBio read mapping
\item "map-hifi" - PacBio HiFi read mapping
\item "sr" - Short-read mapping
\item "asm5" - Long assembly to reference mapping (>95\% identity)
\item "asm10" - Long assembly to reference mapping (>85\% identity)
\item "splice" - Long-read spliced alignment
}}

\item{n_threads}{Number of threads to use for index construction. Default is 3.}
}
\value{
An external pointer to the minimap2 index
}
\description{
Build a minimap2 index from a reference sequence
}
\examples{
\dontrun{
# Build index for Oxford Nanopore reads
idx <- build_index("reference.fa", preset = "map-ont")

# Build index for PacBio reads with 8 threads
idx <- build_index("reference.fa", preset = "map-pb", n_threads = 8)
}
}
