% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2.R
\name{destroy_aligner}
\alias{destroy_aligner}
\title{Clean Up Minimap2 Aligner Object}
\usage{
destroy_aligner(aligner)
}
\arguments{
\item{aligner}{A Minimap2Aligner object}
}
\description{
Clean Up Minimap2 Aligner Object
}
