% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{header}
\alias{header}
\alias{header,AlignmentRecords-method}
\title{Get header from alignment records}
\usage{
header(x)

\S4method{header}{AlignmentRecords}(x)
}
\arguments{
\item{x}{An object with alignment records}
}
\value{
Character vector of header lines
}
\description{
Get header from alignment records
}
