% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2.R
\docType{package}
\name{minimap2}
\alias{minimap2-package}
\alias{minimap2}
\title{R Interface to Minimap2}
\description{
R interface to Minimap2, a versatile sequence alignment program that can be used for short-read alignment, long-read alignment, and short-read assembly-to-reference alignment.
}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
\keyword{internal}
