% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2.R
\name{Minimap2Aligner}
\alias{Minimap2Aligner}
\title{Create a Minimap2 Aligner Object}
\usage{
Minimap2Aligner(
  reference_file,
  preset,
  k = 15,
  w = 10,
  min_cnt = 3,
  min_chain_score = 40,
  bw = 500,
  n_threads = 3
)
}
\arguments{
\item{reference_file}{Path to the reference sequence file (FASTA/FASTQ)}

\item{preset}{Preset parameters for different types of sequences (e.g. "lr:hq", "splice:hq") see details}

\item{k}{K-mer length}

\item{w}{Minimizer window length}

\item{min_cnt}{Minimum number of minimizers}

\item{min_chain_score}{Minimum chaining score}

\item{bw}{Band width}

\item{n_threads}{Number of threads to use}
}
\value{
A Minimap2Aligner object
}
\description{
Create a Minimap2 Aligner Object
}
\details{
The \code{preset} parameter is used to set the preset parameters for different types of sequences.
The available presets are:
\itemize{
\item "map-pb": PacBio CLR genomic reads (~15\% error rate)
\item "map-ont": Oxford Nanopore reads (~20\% error rate)
\item "map-hifi": PacBio HiFi/CCS genomic reads (>99\% accuracy)
\item "lr:hq": Nanopore Q20 genomic reads
\item "sr": Short Illumina reads (~1\% error rate)
\item "splice": Long-read spliced alignment (RNA-seq)
\item "splice:hq": Long-read spliced alignment (RNA-seq) with high quality
\item "asm5": Assembly to reference mapping (<5\% divergence)
\item "asm10": Assembly to reference mapping (<10\% divergence)
\item "ava-pb": PacBio all-vs-all overlap mapping
\item "ava-ont": ONT all-vs-all overlap mapping
}
}
