% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{n_mapped}
\alias{n_mapped}
\alias{n_mapped,AlignmentRecords-method}
\title{Get number of mapped records}
\usage{
n_mapped(x)

\S4method{n_mapped}{AlignmentRecords}(x)
}
\arguments{
\item{x}{An object with alignment records}
}
\value{
Integer number of mapped records
}
\description{
Get number of mapped records
}
