% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{n_records}
\alias{n_records}
\alias{n_records,AlignmentRecords-method}
\title{Get number of records}
\usage{
n_records(x)

\S4method{n_records}{AlignmentRecords}(x)
}
\arguments{
\item{x}{An object with alignment records}
}
\value{
Integer number of records
}
\description{
Get number of records
}
