% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{n_unmapped}
\alias{n_unmapped}
\alias{n_unmapped,AlignmentRecords-method}
\title{Get number of unmapped records}
\usage{
n_unmapped(x)

\S4method{n_unmapped}{AlignmentRecords}(x)
}
\arguments{
\item{x}{An object with alignment records}
}
\value{
Integer number of unmapped records
}
\description{
Get number of unmapped records
}
