% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2.R
\name{print.Minimap2Aligner}
\alias{print.Minimap2Aligner}
\title{Print Minimap2 Aligner Object}
\usage{
\method{print}{Minimap2Aligner}(x, ...)
}
\arguments{
\item{x}{A Minimap2Aligner object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Minimap2 Aligner Object
}
