% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{records}
\alias{records}
\alias{records,AlignmentRecords-method}
\title{Get records from alignment records}
\usage{
records(x)

\S4method{records}{AlignmentRecords}(x)
}
\arguments{
\item{x}{An object with alignment records}
}
\value{
tibble of alignment records
}
\description{
Get records from alignment records
}
