% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{sam_to_records}
\alias{sam_to_records}
\title{Convert SAM output to AlignmentRecords}
\usage{
sam_to_records(sam_lines)
}
\arguments{
\item{sam_lines}{Character vector of SAM format lines}
}
\value{
AlignmentRecords object
}
\description{
Convert SAM output to AlignmentRecords
}
