% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentRecords.R
\name{show,AlignmentRecords-method}
\alias{show,AlignmentRecords-method}
\title{Show method for AlignmentRecords}
\usage{
\S4method{show}{AlignmentRecords}(object)
}
\arguments{
\item{object}{AlignmentRecords object}
}
\description{
Show method for AlignmentRecords
}
