% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_sequences.R
\name{align_sequences}
\alias{align_sequences}
\title{Align sequences using minimap2}
\usage{
align_sequences(
  reference_file,
  query_seqs,
  query_names,
  query_quals,
  preset = "map-ont",
  n_threads = 3
)
}
\arguments{
\item{reference_file}{Path to the reference sequence file (FASTA)}

\item{query_seqs}{Character vector of query sequences}

\item{query_names}{Character vector of query sequence names}

\item{query_quals}{Character vector of quality strings}

\item{preset}{Preset parameters for different sequence types:
\itemize{
\item "map-pb": PacBio CLR genomic reads
\item "map-ont": Oxford Nanopore reads
\item "map-hifi": PacBio HiFi/CCS genomic reads
\item "lr:hq": Nanopore Q20 genomic reads
\item "sr": Short Illumina reads
\item "splice": Long-read spliced alignment
\item "splice:hq": High-quality long-read spliced alignment
\item "asm5": Assembly to reference mapping (<5\% divergence)
\item "asm10": Assembly to reference mapping (<10\% divergence)
}}

\item{n_threads}{Number of threads to use for alignment}
}
\value{
An AlignmentRecords object containing the alignment results
}
\description{
Align sequences using minimap2
}
\examples{
\dontrun{
# Align ONT reads
records <- align_sequences(
  reference_file = "reference.fa",
  query_seqs = c("ACGT...", "TGCA..."),
  query_names = c("read1", "read2"),
  query_quals = c("!!!!", "!!!!"),
  preset = "map-ont"
)

# Check alignment results
n_records(records)
n_mapped(records)
head(records(records))
}

}
